fixed all errors
