+++ src/math_lib/math_lib.cpp
@@ -1,46 +1,12 @@
 #include "math_lib.h"
-#include "cmath"
-
-namespace MathLib
-{
-	bool isEqual(double a, double b, double tolerance = 1.e-8f)
-	{
-		return abs(a - b) <= tolerance;
-	}
-	int GCD(int a, int b) {
-		if (b == 0) return a;
-		return GCD(b, a % b);
-	}
-}
-
 #include <cmath>
+#include <algorithm>
 
 namespace MathLib
 {
-    bool isEqual(double a, double b, double tolerance = 0.000000001)
+    bool isEqual(double a, double b, double tolerance)
     {
         return abs(a - b) <= tolerance;
-    }
- 
-    template <typename T>
-    double sqrt(T a)
-    {
-        if (a == 0)
-            return 0;
-        if (a < 0)
-            throw std::invalid_argument("Cannot get the square root of a negative number");
-        double x = a;
-        double result;
-        int count = 0;
-        while (true)
-        {
-            count++;
-            result = 0.5 * (x + (n / x));
-            if (isEqual(result, x, 0.00001))
-                break;
-            x = result;
-        }
-        return result;
     }
 
     bool isPrime(int n)
@@ -70,5 +36,11 @@ namespace MathLib
         }
         return lcm;
     }
+
+    int GCD(int a, int b)
+    {
+		if (b == 0) return a;
+		return GCD(b, a % b);
+	}
 }
